+++ src/main/java/com/example/review/member/domain/Member.java
@@ -1,0 +1,29 @@
+package com.example.review.member.domain;
+
+import lombok.*;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.Entity;
+import javax.persistence.EntityListeners;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import java.time.LocalDate;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+@Builder
+public class Member {
+    @Id
+    @GeneratedValue
+    Long memberId;
+    private String loginId;
+    private String nickname;
+    private String password;
+    @CreatedDate
+    private LocalDate signUpDate;
+}
