--- src/main/java/com/example/review/post/domain/Member.java
@@ -1,22 +1,0 @@
-package com.example.review.post.domain;
-
-import lombok.*;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Member {
-    @Id
-    @GeneratedValue
-    Long memberId;
-    private String loginId;
-    private String nickname;
-    private String password;
-}
