+++ src/main/java/com/example/review/post/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.example.review.post.repository;
 
-import com.example.review.post.domain.Member;
+import com.example.review.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
